Player Bounce and Rotation
I added the effect where the placer bounces when fall from a height. This is a change from the player not being able to fall off of ledges previously done with raycasts, but we are now looking for the ability to take fall damage. The player also rotates towards where you are walking.